+++ src/components/calcite-popover/calcite-popover.stories.js
@@ -37,7 +37,7 @@ const contentHTML = `
 <div style="padding:12px 16px">
   <b>I am a title!</b> <br>
   <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
-  <calcite-button appearance="inline">I am an inline button</calcite-button>
+  <calcite-link>I am an inline link</calcite-link>
 </div>
 `;